Refactored quality checks so the checks are broken up into individual files, as well as converting _ check classes to ABC's. Official docs on ABC's here: https://docs.python.org/3/library/abc.html. Also a good snippet on why ABC's are useful here: https://stackoverflow.com/questions/3570796/why-use-abstract-base-classes-in-python.
