+++ tsdat/qc/checkers/check_fail_delta.py
@@ -1,0 +1,11 @@
+from .check_delta import CheckDelta
+
+
+class CheckFailDelta(CheckDelta):
+    """------------------------------------------------------------------------------------
+    Checks for deltas between consecutive values larger than 'fail_delta'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "fail_delta"
