+++ tsdat/qc/checkers/check_fail_max.py
@@ -1,0 +1,11 @@
+from .check_max import CheckMax
+
+
+class CheckFailMax(CheckMax):
+    """------------------------------------------------------------------------------------
+    Checks for values greater than 'fail_max'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "fail_max"
