+++ tsdat/qc/checkers/check_fail_min.py
@@ -1,0 +1,11 @@
+from .check_min import CheckMin
+
+
+class CheckFailMin(CheckMin):
+    """------------------------------------------------------------------------------------
+    Checks for values less than 'fail_min'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "fail_min"
