+++ tsdat/qc/checkers/check_fail_range_max.py
@@ -1,0 +1,11 @@
+from .check_max import CheckMax
+
+
+class CheckFailRangeMax(CheckMax):
+    """------------------------------------------------------------------------------------
+    Checks for values greater than 'fail_range'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "fail_range"
