+++ tsdat/qc/checkers/check_fail_range_min.py
@@ -1,0 +1,11 @@
+from .check_min import CheckMin
+
+
+class CheckFailRangeMin(CheckMin):
+    """------------------------------------------------------------------------------------
+    Checks for values less than 'fail_range'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "fail_range"
