+++ tsdat/qc/checkers/check_valid_min.py
@@ -1,0 +1,11 @@
+from .check_min import CheckMin
+
+
+class CheckValidMin(CheckMin):
+    """------------------------------------------------------------------------------------
+    Checks for values less than 'valid_min'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "valid_min"
