+++ tsdat/qc/checkers/check_valid_range_max.py
@@ -1,0 +1,11 @@
+from .check_max import CheckMax
+
+
+class CheckValidRangeMax(CheckMax):
+    """------------------------------------------------------------------------------------
+    Checks for values greater than 'valid_range'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "valid_range"
