+++ tsdat/qc/checkers/check_valid_range_min.py
@@ -1,0 +1,11 @@
+from .check_min import CheckMin
+
+
+class CheckValidRangeMin(CheckMin):
+    """------------------------------------------------------------------------------------
+    Checks for values less than 'valid_range'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "valid_range"
