+++ tsdat/qc/checkers/check_warn_max.py
@@ -1,0 +1,11 @@
+from .check_max import CheckMax
+
+
+class CheckWarnMax(CheckMax):
+    """------------------------------------------------------------------------------------
+    Checks for values greater than 'warn_max'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "warn_max"
