+++ tsdat/qc/checkers/check_warn_min.py
@@ -1,0 +1,11 @@
+from .check_min import CheckMin
+
+
+class CheckWarnMin(CheckMin):
+    """------------------------------------------------------------------------------------
+    Checks for values less than 'warn_min'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "warn_min"
