+++ tsdat/qc/checkers/check_warn_range_max.py
@@ -1,0 +1,11 @@
+from .check_max import CheckMax
+
+
+class CheckWarnRangeMax(CheckMax):
+    """------------------------------------------------------------------------------------
+    Checks for values greater than 'warn_range'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "warn_range"
