+++ tsdat/qc/checkers/check_warn_range_min.py
@@ -1,0 +1,11 @@
+from .check_min import CheckMin
+
+
+class CheckWarnRangeMin(CheckMin):
+    """------------------------------------------------------------------------------------
+    Checks for values less than 'warn_range'.
+
+    ------------------------------------------------------------------------------------
+    """
+
+    attribute_name: str = "warn_range"
